backend/requirements.txt
fastapi
uvicorn
pandas
numpy
akshare
requests
sqlalchemy==2.0.0
pymysql==1.1.0
cryptography==41.0.0
apscheduler==3.10.0
python-dotenv==1.0.0
